+++ CHANGELOG.md
@@ -27,7 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - `Blocks` - ordered list of items to play within the specified duration
     - `Templates` - a generic "day" that consists of blocks scheduled at specific times
     - `Playout Templates` - templates to schedule using the specified criteria. Only one template will be selected each day
-  - Much more to come on this feature as development continues 
+  - Much more to come on this feature as development continues
+- Show chapter markers in movie and episode media info
 
 ### Fixed
 - Fix error loading path replacements when using MySql